# PERF: unstack