--- a/pandas/core/internals/blocks.py
+++ b/pandas/core/internals/blocks.py
@@ -1649,18 +1649,16 @@ class ExtensionBlock(libinternals.Block, EABackedBlock):
         # converting to a 2-D ndarray of objects.
         # Instead, we unstack an ndarray of integer positions, followed by
         # a `take` on the actual values.
-        n_rows = self.shape[-1]
-        dummy_arr = np.arange(n_rows)
 
-        new_values, mask = unstacker.get_new_values(dummy_arr, fill_value=-1)
-        mask = mask.any(0)
-        # TODO: in all tests we have mask.all(); can we rely on that?
+        # Caller is responsible for ensuring self.shape[-1] == len(unstacker.index)
+        new_values, mask = unstacker.arange_result
 
         blocks = [
             # TODO: could cast to object depending on fill_value?
-            self.make_block_same_class(
+            type(self)(
                 self.values.take(indices, allow_fill=allow_fill, fill_value=fill_value),
-                BlockPlacement(place),
+                placement=BlockPlacement(place),
+                ndim=2,
             )
             for indices, place in zip(new_values.T, new_placement)
         ]