```
from asv_bench.benchmarks.reshape import *
self = ReshapeExtensionDtype()
self.setup("Period[s]")

%timeit self.time_unstack_fast('Period[s]')
9.23 ms ± 357 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- master
5.19 ms ± 106 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- PR
```